refactor: move UsbDeviceCtrlTester
